requirements.txt
bandit==1.8.3
flask==3.1.0
gradio==5.25.1
numpy==2.2.4
openai==1.73.0
openai-agents==0.0.9
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3